requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aiomysql==0.0.22
aiosignal==1.2.0
aiosqlite==0.17.0
altgraph==0.17.2
async-timeout==4.0.2
attrs==21.2.0
auto-py-to-exe==2.15.0
Babel==2.9.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
Eel==0.12.4
frozenlist==1.2.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
iso8601==0.1.16
multidict==5.2.0
orjson==3.6.6
pefile==2021.9.3
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyMySQL==0.9.3
pypika-tortoise==0.1.3
python-rapidjson==1.5
pytz==2021.3
pywin32-ctypes==0.2.0
tortoise-orm==0.18.1
typing_extensions==4.0.1
whichcraft==0.6.1
wincertstore==0.2
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
